Turn off all output processing
It turns out that the terminal does a similar translation on the output side. It translates each ("\n") we print into a carriage return followed by a newline ("\r\n").
Turn off OPOST(comes form <tremios.h>. O means it;s an output flag, and I assume POST stands for "post-processing of output"), you'll see that the newline characters we're printing are only moving the cursor down, and not to the left side of the screen.

we need fix that by adding carriage returns to our printf() statements.