A timeout for `read()`
Set a timeout, so that `read()` returns if it doesn't get any input
for a certain amount of time.

=c_cc= field stands for "control characters", an array of bytes that
control various terminal settings.

The =VMIN= value sets the minimum number of bytes of input needed before
read() can return. We set it to 0 so that read() returns as soon as there is any input to
be read. The =VTIME= value sets the maximum amount of time to wait before read()
returns. It is in tenths of a second, so we set it to 1/10 of a second, or 100
milliseconds. If read() times out, it will return 0, which makes sense because its usual return value is the number of bytes read.