Error handling
Most C library functions that fail will set the global errno variable
to indicate what the error was.
perror() looks at the global errno variable and prints a descriptive
error message for it.
It also prints the string given to it before it prints the error
message, which is meant to provide context about what part of your code caused the error.

errno and EAGAIN come from <errno.h>.

tcsetattr(), tcgetattr(), and read() all return -1 on failure, and set
the errno value to indicate the error.