Global state
Our next goal is to get the size of the terminal, so we know how many
rows to draw in editorDrawRows(). But first, let’s set up a global
struct that will contain our editor state, which we’ll use to store
the width and height of the terminal. For now, let’s just put our
orig_termios global into the struct.