Window size, the hard way
=ioctl()= isn’t guaranteed to be able to request the window size on all
systems, so we are going to provide a fallback method of getting the
window size.

The =C= and =B= commands are specifically documented to stop the cursor
from going past the edge of the screen. The reason we don’t use the
=<esc>[999;999H= command is that the documentation doesn’t specify what
happens when you try to move the cursor off-screen.

The =n= command (Device Status Report) can be used to query the terminal
for status information. We want to give it an argument of =6= to ask for
the cursor position. Then we can read the reply from the standard input.