A line viewer
display one line in a file

we have set -std=c11 -Wpedantic, which means that the normal defines
are not available and POSIX function are not declared unless you
explicitly set one of _POSIX_C_SOURCE or _XOPEN_SOURCE. Add #define
_XOPEN_SOURCE 700 (or even 800 on Linux) before the first #include and
all will be well.

that's why compiler complains getline()

we need to define a feature test macro. such as #define
_DEFAULT_SOURCE to enable them.