Multiple lines
To store multiple lines, let’s make E.row an array of erow structs.
It will be a dynamically-allocated array, so we’ll make it a pointer
to erow, and initialize the pointer to NULL.

We want editorAppendRow() to allocate space for a new erow, and then
copy the given string to a new erow at the end of the E.row array.