Snap cursor to end of line
The user is still able to move the cursor past the end of a line,
however. They can do it by moving the cursor to the end of a long
line, then moving it down to the next line, which is shorter. The E.cx
value won’t change, and the cursor will be off to the right of the end of the line it’s now on.