Rendering tabs
The length of a tab is up to the terminal being used and its settings.
We add a second string to the =erow= struct called =render= to render
nonprintable characters tabs, a =^= character followed by another
character, such as =^A= for the =Ctrl-A= character (this is a common way
to display control characters in the terminal).

for now tab is rendering multi (' ')s (no more than 8)