Status bar
The m command (Select Graphic Rendition) causes the text printed after it to
be printed with various possible attributes including bold (1), underscore (4),
blink (5), and inverted colors (7). For example, you could specify all of these
attributes using the command <esc>[1;4;5;7m. An argument of 0 clears all attributes,
and is the default argument, so we use <esc>[m to go back to normal text formatting.

show filename, align it to the left edge of the screen.
show lien number, align it to the right edge of the screen.