Save to disk
open(), O_RDWR, and O_CREAT come from <fcntl.h>.
ftruncate() and close() come from <unistd.h>.

We tell open() we want to create a new file if it doesn’t already
exist (O_CREAT), and we want to open it for reading and
writing (O_RDWR). Because we used the O_CREAT flag, we have to pass an
extra argument containing the mode (the permissions) the new file
should have. 0644 is the standard permissions you usually want for
text files. It gives the owner of the file permission to read and
write the file, and everyone else only gets permission to read the file.

ftruncate() sets the file’s size to the specified length. If the file
is larger than that, it will cut off any data at the end of the file
to make it that length. If the file is shorter, it will add 0 bytes at
the end to make it that length.