Backspacing at the start of a line
you may append current row string to the end of the previous row,
and then delete current row.

Notice that pressing the Delete key at the end of a line works as the
user would expect, joining the current line with the next line. This
is because moving the cursor to the right at the end of a line moves
it to the beginning of the next line. So making the Delete key an
alias for the → key followed by the Backspace key still works.