Restore syntax highlighting after search
We want to restore hl to its previous value after each search. To do
that, we’ll save the original contents of hl in a static variable
named saved_hl in editorFindCallback(), and restore hl to the contents
of saved_hl at the top of the callback.

Notice that the malloc()’d memory is guaranteed to be free()’d,
because when the user closes the search prompt by pressing Enter or
Escape, editorPrompt() calls our callback, giving a chance for hl to
be restored before editorPrompt() finally returns. Also notice that
it’s impossible for saved_hl to get malloc()’d before its old value
gets free()’d, because we always free() it at the top of the function.
And finally, it’s impossible for the user to edit the file between
saving and restoring the hl, so we can safely use saved_hl_line as an
index into E.row. (It’s important to think about these things.)