Nonprintable characters
We use =iscntrl()= to check if the current character is a control
character. If so, we translate it into a printable character by adding
its value to '@' (in ASCII, the capital letters of the alphabet come
after the @ character), or using the '?' character if it’s not in the alphabetic range.